added classic xray module
Check sides checkbox enables / disabled side checks when blocks are being rendered, it is highly recommended this is turned on if you are searching a block that is super common.

If you don't you will probably crash your game. Basically anything the world is made of (stone, end stone, etc) will crash your game if you're not carful. This is because when the module is on and the checkbox is off  every rendered block will render all sides, even sides you cannot see them. This is fine for blocks that are not common like ores, but stone or large amounts of blocks will cause performance issues.